follow & unfollow features improved

+++ src/services/followService.ts
@@ -2,6 +2,7 @@ import { Equal, Repository } from "typeorm";
 import { AppDataSource } from "../data-source";
 import { Follow } from "../entities/Follow";
 import { User } from "../entities/User";
+import ResponseError from "../error/responseError";
 
 export default new (class FollowService {
     private readonly followRepository: Repository<Follow> = AppDataSource.getRepository(Follow);
@@ -27,6 +28,11 @@ export default new (class FollowService {
     }
 
     async follow(follower, following) {
+        const chkFollow = await this.followRepository.countBy({
+            following: Equal(following),
+            follower: Equal(follower),
+        });
+        if (chkFollow) throw new ResponseError(400, "You already follow this User");
         await this.followRepository.save({ following, follower });
         return {
             message: "Follow success",
@@ -41,6 +47,7 @@ export default new (class FollowService {
                 follower: true,
             },
         });
+        if (!getFollow) throw new ResponseError(400, "You already unfollow this User");
         await this.followRepository.delete(getFollow.id);
         return {
             message: "Unfollow success",
